Support for inivation cameras and improved docs (#72)
* Re-introduced support for inivation cameras for pip installs
* Improved 
* Improved documentation on CLI and Python

This PR adds support for *pre-installed* inivation cameras. Meaning, the CLI streaming and the Python `USBInput` will be available **if** the libcaer drivers are available at compile-time.

+++ example/sdl.py
@@ -17,4 +17,4 @@ def create_sdl_surface(*shape):
 
 
 def events_to_bw(events):
-    return events.int() * (255 << 16)
+    return events * (255 << 16)
